Added Categories API
1. API for fetching all categories and fetching category by ID
2. Protected product modification routes based on seller ID
3. Added a dummy helper method for adding categories from the server to firebase

+++ api/v1/products/delete.js
@@ -6,11 +6,17 @@ module.exports = async(req, res) => {
         const id = req.params.id;
 
         const product = await getProductById(id);
+        if(!product){
+            return res.status(403).json({
+                success: false,
+                message: "This action is not possible"
+            });    
+        }
         if(product && product.sellerId!=req.user.id){
-        return res.status(403).json({
-            success: false,
-            message: "You are not allowed to perform this action"
-        })
+            return res.status(403).json({
+                success: false,
+                message: "You are not allowed to perform this action"
+            });
         }
         await deleteProductById(id);
 
@@ -19,11 +25,12 @@ module.exports = async(req, res) => {
             message: "Product deleted successfully"
         });
     } catch (error) {
-        if( error instanceof DatabaseError )
-        return res.status(502).json({
-            success: false,
-            message: error.message
-        });
+        if( error instanceof DatabaseError ){
+            return res.status(502).json({
+                success: false,
+                message: error.message
+            });
+        }
         console.log(error);
         return res.status(500).json({
         success: false,
